Adding MnemonicInput component
